Re-organize oihelper into "util" package
util is a general package for QuickCommand, RobotInspector, and any
other utilities we make.
