+++ reboot2015/src/org/usfirst/frc/team69/robot/subsystems/ContainerGrabber.java
@@ -1,7 +1,7 @@
 package org.usfirst.frc.team69.robot.subsystems;
 
 import org.usfirst.frc.team69.robot.RobotMap;
-import org.usfirst.frc.team69.robot.oihelper.QuickCommand;
+import org.usfirst.frc.team69.util.QuickCommand;
 
 import edu.wpi.first.wpilibj.DoubleSolenoid;
 import edu.wpi.first.wpilibj.DoubleSolenoid.Value;
