+++ reboot2015/src/org/usfirst/frc/team69/util/port/InvalidPortException.java
@@ -1,4 +1,4 @@
-package org.usfirst.frc.team69.robot.oihelper;
+package org.usfirst.frc.team69.util.port;
 
 @SuppressWarnings("serial")
 public class InvalidPortException extends Exception {
